% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CalBoundaries}
\alias{CalBoundaries}
\title{California State Boundaries}
\format{A data frame with 8057 rows and 13 variables:
\describe{
  \item{\code{STATEFP}}{character Unique State Code}
  \item{\code{COUNTYFP}}{character Unique County Code}
  \item{\code{TRACTCE}}{character Unique Census Tract Code}
  \item{\code{GEOID}}{character GEOID}
  \item{\code{NAME}}{character Polygon Name}
  \item{\code{NAMELSAD}}{character Census Tract Name}
  \item{\code{MTFCC}}{character MTFCC}
  \item{\code{FUNCSTAT}}{character FUNCTST}
  \item{\code{ALAND}}{double Land}
  \item{\code{AWATER}}{double Water}
  \item{\code{INTPTLAT}}{character Latitude}
  \item{\code{INTPTLON}}{character Longitude}
  \item{\code{geometry}}{list Polygon Geometry}
}}
\source{
\url{https://data.ca.gov/dataset/ca-geographic-boundaries}
}
\usage{
CalBoundaries
}
\description{
Shapefile for the California state boundaries.
}
\keyword{datasets}
