% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R, R/fire.matrix.R, R/fire.tensor.R
\name{fire}
\alias{fire}
\alias{fire.default}
\alias{fire.data.frame}
\alias{fire.array}
\alias{fire.list}
\alias{fire.matrix}
\alias{fire.tensor}
\title{Fit a FIRe model}
\usage{
fire(X, Y, ...)

\method{fire}{default}(
  X,
  Y,
  ...,
  dat_T,
  scale = TRUE,
  kernels,
  kernels_params,
  kernel_iprior = "cfbm",
  iprior_param = NULL,
  maxiter = 200,
  stop.eps = 1e-05,
  center = FALSE,
  par_init = NULL,
  os_type = "Apple",
  asymptote = TRUE,
  sample_id = 1
)

\method{fire}{data.frame}(X, Y, ...)

\method{fire}{array}(X, Y, ...)

\method{fire}{list}(X, Y, ...)

\method{fire}{matrix}(
  X,
  Y,
  ...,
  dat_T,
  scale = TRUE,
  kernels = list(cfbm),
  kernels_params = list(0.5),
  kernel_iprior = "cfbm",
  iprior_param = NULL,
  maxiter = 200,
  stop.eps = 1e-05,
  center = FALSE,
  par_init = NULL,
  os_type = "Apple",
  asymptote = TRUE
)

\method{fire}{tensor}(
  X,
  Y,
  ...,
  dat_T,
  scale = TRUE,
  kernels,
  kernels_params,
  kernel_iprior = "cfbm",
  iprior_param = NULL,
  maxiter = 200,
  stop.eps = 1e-05,
  center = FALSE,
  par_init = NULL,
  constant = TRUE,
  os_type = "Apple",
  asymptote = TRUE,
  sample_id = 1
)
}
\arguments{
\item{X}{A numeric inputs in matrix, data.frame, array or list form}

\item{Y}{A numeric response vector}

\item{...}{Additional arguments passed to methods}

\item{dat_T}{List of index sets for each mode}

\item{scale}{Logical indicating whether to center the response by subtracting mean(Y)}

\item{kernels}{List of kernel functions for each mode, may refer \code{\link{kernels_fire}} for details}

\item{kernels_params}{List of parameters for each kernel}

\item{kernel_iprior}{Type of I-prior kernel}

\item{iprior_param}{Parameter for I-prior kernel:
\itemize{
\item{\code{"cfbm"}} - Hurst
\item{\code{"rbf"}} - lengthscale
\item{\code{"linear"}} - offset
\item{\code{"poly"}} - degree and offset
}}

\item{maxiter}{Maximum number of EM iterations}

\item{stop.eps}{Convergence tolerance}

\item{center}{Logical indicating whether to center the kernel matrix}

\item{par_init}{Optional list of initial parameter values (lambda, noise)}

\item{os_type}{Operating system type for compatibility ("Apple" or "Windows")}

\item{asymptote}{Logical to use asymptotic initial values}

\item{sample_id}{The sample mode identifier, either the 1st or last mode}

\item{constant}{Logical indicating whether to include the constant kernel term}
}
\value{
An object of class \code{fire_matrix} or \code{fire_tensor}.
The \code{print()} and \code{summary()} methods show the corresponding model information.
}
\description{
A function to perform functional regression using I-priors and Reproducing Kernel Hilbert Space (RKHS).
The FIRe model parameters are estimated by using EM algorithm.
}
\details{
The \code{fire} function is able to take matrix, data.frame, array and list inputs,
the syntax is as per the default S3 method.
}
\section{Methods}{

This generic function has methods for different input types:
\itemize{
  \item \code{\link{fire.matrix}} for matrix/data.frame inputs
  \item \code{\link{fire.tensor}} for array/list inputs
}
}

\examples{
# Matrix input
data(Manure)
mod1 <- fire(X = Manure$absorp[1:5,], Y = Manure$y$DM[1:5],
 dat_T = list(1:700), stop.eps = 2, maxiter = 4)
summary(mod1)

# Array input
data(Housing)
dat_T <- list(T1 = 1:4, T2 = 1:9)
mod2 <- fire(X = Housing$X[1:5,,], Y = Housing$y[1:5,2],
kernels = list(kronecker_delta, kronecker_delta),
kernels_params = list(NA, NA),
dat_T = dat_T, stop.eps = 2, maxiter = 4)
summary(mod2)
}
\seealso{
\code{\link{kernels_fire}}, \code{\link{Manure}}, \code{\link{Housing}}
}
