% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Firestat}
\alias{Firestat}
\title{Firestations}
\format{A data frame with 860 rows and 10 variables:
\describe{
  \item{\code{zip}}{character Zipcode}
  \item{\code{FDID}}{integer Unique Firestation ID}
  \item{\code{name}}{integer Firestation Dept Name}
  \item{\code{address}}{integer Address of firestation}
  \item{\code{city}}{integer City of firestation}
  \item{\code{state}}{integer State of firestation}
  \item{\code{location}}{character Compiled address}
  \item{\code{latitude}}{double Latitude of firestation}
  \item{\code{longitude}}{double Longitude of firestation}
  \item{\code{geometry}}{list Point geometry of firestation}
}}
\source{
\url{https://apps.usfa.fema.gov/registry/download}
}
\usage{
Firestat
}
\description{
This data includes the point locations for all the firestations in California.
}
\keyword{datasets}
