% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_fitted.R
\name{fitted.fire}
\alias{fitted.fire}
\alias{fitted}
\alias{fitted.fire_matrix}
\alias{fitted.fire_tensor}
\alias{print.fire_fitted}
\title{Obtain fitted values of FIRE model}
\usage{
fitted(object, ...)

\method{fitted}{fire}(object, ...)

\method{fitted}{fire_matrix}(object, ...)

\method{fitted}{fire_tensor}(object, ...)

\method{print}{fire_fitted}(x, ...)
}
\arguments{
\item{object}{A \code{fire_matrix} or \code{fire_tensor} object}

\item{...}{Not used}

\item{x}{A \code{fire_fitted} object to print}
}
\value{
A list of class \code{fire_fitted} containing the fitted values,
training RMSE, residuals and intercept.

The returned object has a \code{print} method that displays:
\itemize{
  \item Training RMSE
  \item Intercept value
  \item First few fitted values
}
}
\description{
Obtain fitted values of FIRE model
}
\examples{
data(Manure)
mod <- fire(X = Manure$absorp[1:5,], Y = Manure$y$DM[1:5],
dat_T = list(1:700),, stop.eps = 2, maxiter = 4)
fitted(mod)

}
\seealso{
\code{\link{fire}}
}
