% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_fitted.R
\name{fitted.fire_matrix}
\alias{fitted.fire_matrix}
\title{Fitted Values for \code{fire_matrix} Objects}
\usage{
\method{fitted}{fire_matrix}(object, ...)
}
\arguments{
\item{object}{A model object of class \code{fire_matrix}.}

\item{...}{Not used.}
}
\description{
Computes fitted values for FIRE models with matrix input data.
}
\examples{
data(Manure)
mod <- fire(X = Manure$absorp[1:5, ], Y = Manure$y$DM[1:5],
            dat_T = list(1:700), stop.eps = 2, maxiter = 4)
fitted(mod)

}
