% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_fitted.R
\name{fitted}
\alias{fitted}
\title{Fitted Values for FIRE Models}
\usage{
fitted(object, ...)
}
\arguments{
\item{object}{A model object of class \code{fire_matrix} or \code{fire_tensor}.}

\item{...}{Not used.}
}
\value{
A list of class `fire_fitted` containing:
\itemize{
  \item `yhat`: Fitted values
  \item `rmse`: Training RMSE
  \item `residuals`: Model residuals
  \item `intercept`: Intercept value
  \item `model`: Reference to the original model object
}
}
\description{
Computes fitted values for FIRE models of class \code{fire_matrix} or \code{fire_tensor}.
}
\examples{
data(Manure)
mod <- fire(X = Manure$absorp[1:5,], Y = Manure$y$DM[1:5],
dat_T = list(1:700),, stop.eps = 2, maxiter = 4)
fitted(mod)

}
\seealso{
\code{\link{fire}}
}
