% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprior_get_gram.R
\name{iprior_get_gram}
\alias{iprior_get_gram}
\alias{rbf_rkhs_kron}
\alias{rbf_rkhs_kron_cross}
\alias{cfbm_rkhs_kron}
\alias{cfbm_rkhs_kron_cross}
\title{Compute Gram Matrices for I-prior Models}
\usage{
rbf_rkhs_kron(nmat, lengthscale = 1)

rbf_rkhs_kron_cross(nmat_cross, lengthscale = 1)

cfbm_rkhs_kron(nmat, Hurst = 0.5)

cfbm_rkhs_kron_cross(nmat, nmat_cross, Hurst = 0.5)
}
\arguments{
\item{nmat}{RKHS norm matrix of training data}

\item{lengthscale}{Lengthscale parameter for RBF kernel}

\item{nmat_cross}{Matrix of RKHS norm between training and test data}

\item{Hurst}{Hurst coefficient for cfbm kernel}
}
\value{
A Gram matrix of appropriate dimensions:
\itemize{
  \item For \code{rbf_rkhs_kron} and \code{cfbm_rkhs_kron}: Square training Gram matrix
  \item For \code{rbf_rkhs_kron_cross} and \code{cfbm_rkhs_kron_cross}: Cross Gram matrix between test and training data
}
}
\description{
Functions to compute various Gram matrices for I-prior models using RKHS norms.
}
\examples{
# Using internal functions to compute pairwise RKHS norm
G <- list(matrix(c(2,1,1,2),2), matrix(c(3,1,1,3),2))
X <- list(matrix(rnorm(4),2), matrix(rnorm(4),2))
nmat <- fire:::Kronecker_norm_mat(X, G, alpha=c(0.5,0.5))
ncross <- fire:::Kronecker_norm_cross(X, X[1], G, alpha=c(0.5,0.5))

# Get Gram matrices
K_rbf <- rbf_rkhs_kron(nmat)
K_cfbm <- cfbm_rkhs_kron(nmat)
K_rbf_cross <- rbf_rkhs_kron_cross(ncross)
K_cfbm_cross <- cfbm_rkhs_kron_cross(nmat, ncross)

}
