\name{sample_label}
\docType{data}
\alias{sample_label}
\title{Labels of preprocessed 293T–Jurkat scRNA-seq dataset}
\description{
  This contains labels for 1580 observations. There are two different clusters,
  one is abundant (denoted by 1) and the other is rare (denoted by 2).
  Rare cluster consists of Jurkat cells, which form 2.5\% of the total population.
  The abundant cluster consists of 293T cells.
}
\usage{data(sample_label)}
\format{A vector containing 1580 observations.}
\source{10xgenomics, https://support.10xgenomics.com/single-cell-gene-expression/datasets}
\references{
  Zheng, G. X. et al.
  Massively parallel digital transcriptional profiling of single cells.
  \emph{Nat. Commun.} 8, 14049 (2017).
}
\keyword{datasets}
