% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Manure}
\alias{Manure}
\title{French Manure Data}
\format{
A list containing two components:
\describe{
  \item{\code{absorp}}{A matrix of NIR absorbance spectra (332 samples × 700 wavelengths). Wavelength range: 1100-2500 nm (2 nm resolution)}
  \item{\code{y}}{A matrix of chemical measurements (332 samples × 3 variables)}
  \itemize{
  \item \code{DM}: Dry matter content (\% of wet weight)
  \item \code{NH4}: Total ammonium nitrogen (\% of wet weight)
  \item \code{N}: Total nitrogen (\% of wet weight)
}
}
}
\usage{
Manure
}
\description{
Combined near-infrared spectra and chemical composition measurements of cattle and poultry manure samples.
}
\examples{
data(Manure)
}
\references{
Gogé, F., Thuriès, L., Fouad, Y., Damay, N., Davrieux, F., Moussard, G., ... & Morvan, T. (2021).
Dataset of chemical and near-infrared spectroscopy measurements of fresh and dried poultry and cattle manure.
\emph{Data in Brief}, 34, 106647. \doi{10.1016/j.dib.2020.106647}
}
\keyword{datasets}
