% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{PeMS}
\alias{PeMS}
\title{PeMS Highway Traffic Data}
\format{
A 3-way tensor (array) with dimensions:
\describe{
  \item{X}{A 3-way tensor (array) with dimensions:
    \itemize{
      \item{Hours: 24 hours of daily measurements}
      \item{Days: 43 days of observations (weekdays of May-June 2012)}
      \item{Sensors: 228 sensor stations in District 7}
    }
  }
  \item{y}{A vector containing the average traffic speed of 228 sensors
           during the morning peak hours (6-9am) on the 44th day}
}
}
\source{
Downloaded on 2025-07-19 and processed by package author.
}
\usage{
PeMS
}
\description{
Hourly traffic speed measurements from California highway sensors.
}
\details{
Processed from the original PeMSD7 dataset containing raw 5-minutes measurements aggregated to hourly intervals
}
\examples{
data(PeMS)
}
\references{
B. Yu, H. Yin, and Z. Zhu, "Spatio-temporal graph convolutional networks:
A deep learning framework for traffic forecasting," in Proc. IJCAI, 2018, pp. 3634–3640.
}
\keyword{datasets}
