% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_plot.R
\name{plot.fire}
\alias{plot.fire}
\alias{plot}
\title{Plot Method for FIRE Models}
\usage{
plot(object, ...)
}
\arguments{
\item{...}{Not used}

\item{x}{A \code{fire_fitted} or \code{fire_prediction} object}
}
\value{
Returns a plot object.
  The exact return depends on the input:
  \itemize{
    \item For \code{fire_fitted} object: Returns a single \code{ggplot} object showing residuals vs fitted values
    \item For \code{fire_prediction} object:
      \itemize{
        \item When test data is available: Returns a \code{gridExtra}-arranged plot containing both
              residuals vs predicted and actual vs predicted plots
        \item When no test data is available: Returns a single \code{ggplot} object showing
              the distribution of predicted values with density overlay
      }
  }
}
\description{
The type of plot produced depends on
the class of the model object \code{fire_fitted} or \code{fire_prediction}
}
\examples{
data(Manure)
idx <- 1:5
mod <- fire(X = Manure$absorp[idx,], Y = Manure$y$DM[idx],
 dat_T = list(1:700), stop.eps = 2, maxiter = 4)

Yfitted = fitted(mod)
plot(Yfitted)

Ypred = predict(mod, newdata = Manure$absorp[idx+10,],
Ynew = Manure$y$DM[idx+10])
plot(Ypred)
}
\seealso{
\code{\link{fire}}, \code{\link{fitted.fire}}, \code{\link{predict.fire}}
}
