% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_plot.R
\name{plot.fire_prediction}
\alias{plot.fire_prediction}
\title{Plot for FIRE Predictions}
\usage{
\method{plot}{fire_prediction}(x, ...)
}
\arguments{
\item{x}{A \code{fire_prediction} object}

\item{...}{Not used}
}
\description{
Generates residuals vs predicted and actual vs predicted plots when test response is provide.
Otherwise, generates the distribution of predicted values with density overlay.
}
\examples{
data(Manure)
mod <- fire(X = Manure$absorp[1:5,], Y = Manure$y$DM[1:5],
            dat_T = list(1:700), stop.eps = 2, maxiter = 4)
pred <- predict(mod, newdata = Manure$absorp[6:10,],
                Ynew = Manure$y$DM[6:10])
plot(pred)

}
