% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_predict.R
\name{predict.fire_matrix}
\alias{predict.fire_matrix}
\title{Predict for \code{fire_matrix} Objects}
\usage{
\method{predict}{fire_matrix}(object, newdata, Ynew = NULL, ...)
}
\arguments{
\item{object}{A model object of class \code{fire_matrix}.}

\item{newdata}{New data for prediction.}

\item{Ynew}{Optional true response values for test set evaluation.}

\item{...}{Not used.}
}
\description{
Obtain predictions from FIRE models with matrix input data.
}
\examples{
data(Manure)
mod <- fire(X = Manure$absorp[1:5, ], Y = Manure$y$DM[1:5],
            dat_T = list(1:700), stop.eps = 2, maxiter = 4)
predict(mod, newdata = Manure$absorp[6:10, ])

}
