% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_predict.R
\name{predict.fire_tensor}
\alias{predict.fire_tensor}
\title{Predict for \code{fire_tensor} Objects}
\usage{
\method{predict}{fire_tensor}(object, newdata, Ynew = NULL, ...)
}
\arguments{
\item{object}{A model object of class \code{fire_tensor}.}

\item{newdata}{New data for prediction.}

\item{Ynew}{Optional true response values for test set evaluation.}

\item{...}{Not used.}
}
\description{
Obtain predictions from FIRE models with tensor input data.
}
\examples{
data(Housing)
dat_T <- list(T1 = 1:4, T2 = 1:9)
mod <- fire(X = Housing$X[1:5,,], Y = Housing$y[1:5,2],
            kernels = list(kronecker_delta, kronecker_delta),
            kernels_params = list(NA, NA),
            dat_T = dat_T, stop.eps = 2, maxiter = 4)
predict(mod, newdata = Housing$X[6:10,,])

}
