% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_print.R
\name{print.fire}
\alias{print.fire}
\alias{print.fire_matrix}
\alias{print.fire_tensor}
\title{Print method for FIRE models}
\usage{
\method{print}{fire_matrix}(x, ...)

\method{print}{fire_tensor}(x, ...)
}
\arguments{
\item{x}{A \code{fire_matrix} or \code{fire_tensor} object}

\item{...}{Not used}
}
\value{
The input object (invisibly) for piping. Prints to console:
\itemize{
\item{Basic model information}
\item{Convergence status}
\item{Dimensions}
\item{Estimated hyperparameters}
\item{Marginal log-likelihood}
}
}
\description{
Compact display of \code{fire_matrix} or \code{fire_tensor} objects.
}
\examples{
data(Manure)
mod <- fire(X = Manure$absorp[1:10,], Y = Manure$y$DM[1:10],
 dat_T = list(1:700), stop.eps = 2, maxiter = 4)
print(mod)

}
\seealso{
\code{\link{fire}}
}
