% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_fitted.R
\name{print.fire_fitted}
\alias{print.fire_fitted}
\title{Print FIRE Fitted Values}
\usage{
\method{print}{fire_fitted}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fire_fitted}.}

\item{...}{Not used.}
}
\value{
Invisibly returns the input object.
}
\description{
Prints summarized results for \code{fire_fitted} objects.
}
\seealso{
\code{\link{fitted.fire}}
}
