% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_print.R
\name{print.fire_matrix}
\alias{print.fire_matrix}
\title{Print for \code{fire_matrix} Objects}
\usage{
\method{print}{fire_matrix}(x, ...)
}
\arguments{
\item{x}{A model object of class \code{fire_matrix}.}

\item{...}{Not used.}
}
\description{
Compact display of matrix-input FIRE models.
}
\examples{
data(Manure)
mod <- fire(X = Manure$absorp[1:10,], Y = Manure$y$DM[1:10],
            dat_T = list(1:700), stop.eps = 2, maxiter = 4)
print(mod)  # Or simply type 'mod'

}
