% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_print.R
\name{print.fire_tensor}
\alias{print.fire_tensor}
\title{Print for \code{fire_tensor} Objects}
\usage{
\method{print}{fire_tensor}(x, ...)
}
\arguments{
\item{x}{A model object of class \code{fire_tensor}.}

\item{...}{Not used.}
}
\description{
Compact display of tensor-input FIRE models.
}
\examples{
data(Housing)
dat_T <- list(T1 = 1:4, T2 = 1:9)
mod <- fire(X = Housing$X[1:5,,], Y = Housing$y[1:5,2],
            kernels = list(kronecker_delta, kronecker_delta),
            kernels_params = list(NA, NA),
            dat_T = dat_T, stop.eps = 2, maxiter = 4)
print(mod)

}
