% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_print.R
\name{print}
\alias{print}
\title{Print FIRE Model Objects}
\usage{
print(x, ...)
}
\arguments{
\item{x}{A model object of class \code{fire_matrix} or \code{fire_tensor}.}

\item{...}{Not used.}
}
\value{
The input object (invisibly) for piping. Prints to console:
\itemize{
  \item Model type (matrix/tensor)
  \item Convergence status
  \item Data dimensions
  \item Marginal log-likelihood
  \item Estimated hyperparameters
}
}
\description{
Compact display of FIRE model objects.
}
\seealso{
\code{\link{fire}}
}
