% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_summary.R
\name{print.summary.fire}
\alias{print.summary.fire}
\alias{print.summary.fire_matrix}
\alias{print.summary.fire_tensor}
\title{Print Summary for FIRE Models}
\usage{
\method{print}{summary.fire_matrix}(x, ...)

\method{print}{summary.fire_tensor}(x, ...)
}
\arguments{
\item{x}{A \code{summary.fire} object (from either matrix or tensor input).}

\item{...}{Not used.}
}
\value{
Invisibly returns the input object.
}
\description{
Displays formatted summary output for FIRE models.
}
\examples{
# Print matrix model summary
data(Manure)
mod <- fire(X = Manure$absorp[1:10,], Y = Manure$y$DM[1:10],
            dat_T = list(1:700), control = list(stop.eps = 2, maxiter = 4))
print(summary(mod))

# Print tensor model summary
data(Housing)
dat_T <- list(T1 = 1:4, T2 = 1:9)
mod <- fire(X = Housing$X[1:5,,], Y = Housing$y[1:5,2],
            kernels = list(kronecker_delta, kronecker_delta),
            kernels_params = list(NA, NA),
            dat_T = dat_T, control = list(stop.eps = 2, maxiter = 4))
print(summary(mod))
}
