% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_summary.R
\name{print.summary.fire_matrix}
\alias{print.summary.fire_matrix}
\title{Print Summary for \code{fire_matrix} Objects}
\usage{
\method{print}{summary.fire_matrix}(x, ...)
}
\arguments{
\item{x}{A \code{summary.fire_matrix} object.}

\item{...}{Not used.}
}
\value{
Invisibly returns the input object.
}
\description{
Displays formatted summary output for matrix-based FIRE models.
}
