% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_summary.R
\name{print.summary.fire_tensor}
\alias{print.summary.fire_tensor}
\title{Print Summary for \code{fire_tensor} Objects}
\usage{
\method{print}{summary.fire_tensor}(x, ...)
}
\arguments{
\item{x}{A \code{summary.fire_tensor} object.}

\item{...}{Not used.}
}
\value{
Invisibly returns the input object.
}
\description{
Displays formatted summary output for tensor-based FIRE models.
}
