% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_plot.R
\name{station_plot}
\alias{station_plot}
\title{Plotting the Firestation Locations}
\usage{
station_plot(maintitle, color)
}
\arguments{
\item{maintitle}{title for plot (NULL for no title)}

\item{color}{color of dots}
}
\value{
California polygon with overlayed firestation point data.
}
\description{
\code{station_plot()} returns a map of firestation locations in California
}
\details{
This function allows you to easily plot firestation locations in California onto a map with some basic customizations.
}
\examples{
station_plot("Firestations", 'red')

}
