% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_summary.R
\name{summary.fire}
\alias{summary.fire}
\alias{summary.fire_matrix}
\alias{summary.fire_tensor}
\title{Summary Method for FIRE Models}
\usage{
\method{summary}{fire_matrix}(object, ...)

\method{summary}{fire_tensor}(object, ...)
}
\arguments{
\item{object}{A model object of class \code{fire_matrix} or \code{fire_tensor}.}

\item{...}{Not used.}
}
\value{
An object of class \code{summary.fire} containing:
\itemize{
  \item Estimated parameters (lambda/alpha, noise, intercept)
  \item Convergence information
  \item Model fit statistics (log-likelihood, sample size)
  \item Kernel specifications
  \item Computation timing
}
}
\description{
Detailed summary of FIRE models for either matrix or tensor input data.
}
\examples{
# For matrix input
data(Manure)
mod <- fire(X = Manure$absorp[1:10,], Y = Manure$y$DM[1:10],
            dat_T = list(1:700), control = list(stop.eps = 2, maxiter = 4))
summary(mod)

# For tensor input
data(Housing)
dat_T <- list(T1 = 1:4, T2 = 1:9)
mod <- fire(X = Housing$X[1:5,,], Y = Housing$y[1:5,2],
            kernels = list(kronecker_delta, kronecker_delta),
            kernels_params = list(NA, NA),
            dat_T = dat_T, control = list(stop.eps = 2, maxiter = 4))
summary(mod)
}
\seealso{
\code{\link{fire}}
}
