% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_summary.R
\name{summary.fire_matrix}
\alias{summary.fire_matrix}
\title{Summary for \code{fire_matrix} Objects}
\usage{
\method{summary}{fire_matrix}(object, ...)
}
\arguments{
\item{object}{A model object of class \code{fire_matrix}.}

\item{...}{Not used.}
}
\description{
Detailed summary of FIRE models with matrix input data.
}
\examples{
data(Manure)
mod <- fire(X = Manure$absorp[1:10,], Y = Manure$y$DM[1:10],
            dat_T = list(1:700), stop.eps = 2, maxiter = 4)
summary(mod)

}
