% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_summary.R
\name{summary.fire_tensor}
\alias{summary.fire_tensor}
\title{Summary for \code{fire_tensor} Objects}
\usage{
\method{summary}{fire_tensor}(object, ...)
}
\arguments{
\item{object}{A model object of class \code{fire_tensor}.}

\item{...}{Not used.}
}
\description{
Detailed summary of FIRE models with tensor input data.
}
\examples{
data(Housing)
dat_T <- list(T1 = 1:4, T2 = 1:9)
mod <- fire(X = Housing$X[1:5,,], Y = Housing$y[1:5,2],
            kernels = list(kronecker_delta, kronecker_delta),
            kernels_params = list(NA, NA),
            dat_T = dat_T, stop.eps = 2, maxiter = 4)
summary(mod)

}
