% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_summary.R
\name{summary}
\alias{summary}
\title{Summary of FIRE Models}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{A model object of class \code{fire_matrix} or \code{fire_tensor}.}

\item{...}{Not used.}
}
\value{
An object of class `summary.fire_matrix` or `summary.fire_tensor` containing:
\itemize{
  \item Estimated parameters (lambda/alpha, noise, intercept)
  \item Convergence information
  \item Model fit statistics (log-likelihood, sample size)
  \item Kernel specifications
  \item Computation timing
}
}
\description{
Generates comprehensive summaries for FIRE models of class \code{fire_matrix} or \code{fire_tensor}.
}
\seealso{
\code{\link{fire}}
}
