% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_classification.R
\name{utils_classification}
\alias{utils_classification}
\alias{to_dummy_vector}
\alias{onehot_argmax}
\alias{onehot_to_labels}
\title{Classification Utility Functions}
\usage{
to_dummy_vector(x, levels = NULL)

onehot_argmax(vec, c)

onehot_to_labels(onehot, levels)
}
\arguments{
\item{x}{Categorical vector (character, numeric, or factor)}

\item{levels}{Vector of category levels.}

\item{vec}{Numeric vector of length \eqn{n \times c}, where \eqn{n} is the number
of observations and \eqn{c} is the number of categories.}

\item{c}{Integer giving the number of categories.}

\item{onehot}{A numeric vector of length \eqn{n \times c}, the output from
\code{onehot_argmax}.}
}
\value{
\code{to_dummy_vector}: A numeric vector of length \eqn{n \times} \code{length(levels)}

\code{onehot_argmax}: A numeric vector of length \eqn{n \times c} containing a one-hot encoding
where, within each block of size \eqn{c}, the maximum element is set to 1 and all others to 0.

\code{onehot_to_labels}: A vector of length \eqn{n} containing the
  predicted category labels corresponding to the one-hot encoding.
}
\description{
Collection of helper functions for classification problems
}
\examples{
to_dummy_vector(c("A","B","A"), levels = c("A","B","C"))
to_dummy_vector(c(0,1,0,9), levels = 0:9)
set.seed(1)
v <- runif(12) # n=4, c=3
r <- onehot_argmax(v, c = 3)
r
onehot_to_labels(r, levels = c("A", "B", "C"))
}
