% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tensor.R
\name{utils_tensor}
\alias{utils_tensor}
\alias{tensor_sample}
\alias{vectorize_tensor}
\alias{unfolding}
\alias{dat_unfolding}
\alias{reshape_tensor}
\title{Tensor Utility Functions}
\usage{
tensor_sample(X, sample_id = 1)

vectorize_tensor(X, Index)

reshape_tensor(Mat, dim, N, Index)

unfolding(X, n, mode = TRUE)

dat_unfolding(X, sample_id = 1, mode = TRUE)
}
\arguments{
\item{X}{Input tensor}

\item{sample_id}{Sampling mode (1st or last mode)}

\item{Index}{Matrix of indices}

\item{Mat}{A matrix with N rows, each containing values to be assigned to the corresponding positions in the output tensor}

\item{dim}{Dimensions of output tensor}

\item{N}{Sample size}

\item{n}{Unfolding mode}

\item{mode}{Logical to print status message}
}
\value{
For \code{tensor_sample}: List of tensor samples

For `vectorize_tensor`: Vectorized tensor elements

For `reshape_tensor`: A tensor with specified dimensions and the 1st mode corresponds to sampling mode

For `unfolding`: Unfolded matrix

For `dat_unfolding`: List of mode-n unfolded matrices
}
\description{
Collection of helper functions for tensor operations
}
\examples{
# tensor_sample()
X <- array(1:24, dim = c(3,2,4))
tensor_sample(X, sample_id = 1)
# vectorize_tensor()
X <- array(1:8, dim = c(2,2,2))
idx <- expand.grid(T1 = 1:2, T2 = 1:2, T3 = 1:2)
vectorize_tensor(X, idx)
idx <- cbind(rep(1:2,each = 4), rep(c(1,1,2,2),2), rep(1:2, 4))
vectorize_tensor(X, idx)
# reshape_tensor()
N <- 2
Mat <- matrix(1:12, nrow = N)
dim <- c(2,3)
idx <- expand.grid(1:2, 1:3)
reshape_tensor(Mat, dim, N, idx)
# unfolding()
X <- array(1:8, dim = c(2,2,2))
unfolding(X, 1)
# dat_unfolding()
X <- array(1:24, dim = c(3,2,4))
dat_unfolding(X, sample_id = 1)
}
