% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_plot.R
\name{weather_plot}
\alias{weather_plot}
\title{Plotting the Weather Data}
\usage{
weather_plot(maintitle, attribute, legendtitle, palette)
}
\arguments{
\item{maintitle}{title for plot (NULL for no title)}

\item{attribute}{"Avg_Precip" or "Avg_Temp"}

\item{legendtitle}{title for legend}

\item{palette}{ColorBrewer palette name}
}
\value{
California polygon with overlayed weather attribute data colored by intensity.
}
\description{
\code{weather_plot()} returns a map of point data on a California polygon with dots colored by intensity of attribute.
}
\details{
This function allows you to easily plot the provided weather data onto a map with some basic customizations.
}
\examples{
weather_plot("Avg Temp", "Avg_Temp", "Temp", "Reds")

}
