% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WeatherGeo}
\alias{WeatherGeo}
\title{Weather Data}
\format{A data frame with 336 rows and 5 variables:
\describe{
  \item{\code{STATION}}{integer Unique Weather Station Id}
  \item{\code{NAME}}{integer Station Name}
  \item{\code{Avg_Precip}}{double Average Precipiation for the Year in inches}
  \item{\code{Avg_Temp}}{double Average Temperature for the Year in degrees Fahrenheit}
  \item{\code{geometry}}{list Point Geometry for the Station}
}}
\source{
\url{https://www.ncdc.noaa.gov/cdo-web/datasets}
}
\usage{
WeatherGeo
}
\description{
Temperature and precipitation data from the NOAA weather stations around California. This is the "annual normals" dataset which includes data from 1981-2010 from all the weather stations in California compiled to create an "annual normal".
}
\keyword{datasets}
