% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\name{fire}
\alias{fire}
\title{Convert a function into a CLI}
\usage{
fire(f, ..., run_interactive = FALSE)
}
\arguments{
\item{f}{A function; the function to be called with CLI.}

\item{...}{Additional arguments to be passed to the function.}

\item{run_interactive}{TRUE / FALSE; whether to run the function also in
interactive mode. Default to FALSE.}
}
\description{
Convert a function into a CLI
}
\examples{
# # add.R
# add <- function(x, y) x + y
# fire::fire(add, x = 3, y = 4, TRUE)  # Generate result 1

# # At terminal
# Rscript add.R --x=1 --y=2  # Reuse function to generate result 2
# Rscript add.R help  # Show what arguments are required

}
