% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{package}
\name{fire}
\alias{fire}
\title{fire: A package for mapping fire related data.}
\description{
The fire package provides cleaned datasets related to understanding fire risk and simple functions to map these data.
}
\section{Datasets}{

\describe{
  \item{\code{Firestat}}{Firestations across California}
  \item{\code{WeatherGeo}}{Average temperature and precipitation sample data from NOAA weather stations across California}
  \item{\code{CalBoundaries}}{California state boundary polygon shapefile}
}
}

\section{Functions}{

\describe{
  \item{\code{stations_plot}}{maps the firestations across California}
  \item{\code{weather_plot}}{maps the attribute data provided from the NOAA weather stations across California creating dots with choropleth coloring}
}
}

