% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R, R/fire.matrix.R, R/fire.tensor.R
\name{fire}
\alias{fire}
\alias{fire.default}
\alias{fire.data.frame}
\alias{fire.array}
\alias{fire.list}
\alias{fire.matrix}
\alias{fire.tensor}
\title{Fit a FIRE Model}
\usage{
fire(
  X,
  Y,
  dat_T,
  kernels,
  kernels_params,
  kernel_iprior = "cfbm",
  iprior_param = NULL,
  control = list(),
  ...
)

\method{fire}{default}(
  X,
  Y,
  dat_T,
  kernels,
  kernels_params,
  kernel_iprior = "cfbm",
  iprior_param = NULL,
  control = list(),
  ...
)

\method{fire}{data.frame}(
  X,
  Y,
  dat_T,
  kernels = list(cfbm),
  kernels_params = list(0.5),
  kernel_iprior = "cfbm",
  iprior_param = NULL,
  control = list(),
  ...
)

\method{fire}{array}(
  X,
  Y,
  dat_T,
  kernels,
  kernels_params,
  kernel_iprior = "cfbm",
  iprior_param = NULL,
  control = list(),
  ...
)

\method{fire}{list}(
  X,
  Y,
  dat_T,
  kernels,
  kernels_params,
  kernel_iprior = "cfbm",
  iprior_param = NULL,
  control = list(),
  ...
)

\method{fire}{matrix}(
  X,
  Y,
  dat_T,
  kernels = list(cfbm),
  kernels_params = list(0.5),
  kernel_iprior = "cfbm",
  iprior_param = NULL,
  control = list(),
  ...
)

\method{fire}{tensor}(
  X,
  Y,
  dat_T,
  kernels,
  kernels_params,
  kernel_iprior = "cfbm",
  iprior_param = NULL,
  control = list(),
  ...
)
}
\arguments{
\item{X}{A numeric input in \code{matrix}, \code{data.frame}, \code{array} or \code{list}}

\item{Y}{A numeric response vector}

\item{dat_T}{List of index sets for each mode}

\item{kernels}{List of kernel functions for each mode (see \code{\link{kernels_fire}})}

\item{kernels_params}{List of parameters for each kernel}

\item{kernel_iprior}{Type of I-prior kernel}

\item{iprior_param}{Parameter for I-prior kernel:
\itemize{
  \item \code{"cfbm"} - Hurst coefficient (default 0.5)
  \item \code{"rbf"} - lengthscale (default 1)
  \item \code{"linear"} - offset (default 0)
  \item \code{"poly"} - degree and offset (default c(2, mean(Y)))
}}

\item{control}{A list of control parameters (see Details)}

\item{...}{Additional arguments passed to methods}
}
\description{
A function to perform functional regression using I-priors with Reproducing Kernel Hilbert Space (RKHS) norm.
The FIRE model parameters are estimated using an EM algorithm.
}
\details{
The \code{control} argument can include the following parameters:
\itemize{
  \item{\code{scale}: Logical indicating whether to center the response (default TRUE)}
  \item{\code{maxiter}: Maximum number of EM iterations (default 200)}
  \item{\code{stop.eps}: Convergence tolerance (default 1e-3)}
  \item{\code{constant_g}: Logical indicating whether to include constant kernel term in g (default TRUE)}
  \item{\code{constant_h}: Logical indicating whether to include constant kernel term in h (default FALSE)}
  \item{\code{center}: Logical indicating whether to center the kernel matrix (default FALSE)}
  \item{\code{par_init}: Optional list of initial parameter values (lambda, noise)}
  \item{\code{os_type}: Operating system type for compatibility ("Apple" or "Windows", default "Apple")}
  \item{\code{cores}: Number of cores for parallel computation (default: detectCores() - 1)}
  \item{\code{asymptote}: Logical to use asymptotic initial values (default TRUE)}
  \item{\code{sample_id}: Which mode contains samples (default 1)}
  \item{\code{epsilon}: Small positive constant in initialisation of EM algorithm (default 1)}
}
}
\section{Methods}{

This generic function has methods for different input types:
\itemize{
  \item \code{\link{fire.matrix}} for \code{matrix} or \code{data.frame} inputs
  \item \code{\link{fire.tensor}} for \code{array} or \code{list} inputs
}
}

\examples{
# Matrix input
data(Manure)
mod1 <- fire(X = Manure$absorp[1:5,], Y = Manure$y$DM[1:5],
 dat_T = list(1:700), control = list(stop.eps = 2, maxiter = 4))
summary(mod1)

# Array input
data(Housing)
dat_T <- list(T1 = 1:4, T2 = 1:9)
mod2 <- fire(X = Housing$X[1:5,,], Y = Housing$y[1:5,2],
kernels = list(kronecker_delta, kronecker_delta),
kernels_params = list(NA, NA),
dat_T = dat_T, control = list(stop.eps = 2, maxiter = 4))
summary(mod2)
}
\seealso{
\code{\link{kernels_fire}}, \code{\link{Manure}}, \code{\link{Housing}}
}
