% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_fitted.R
\name{fitted.fire}
\alias{fitted.fire}
\alias{fitted.fire_matrix}
\alias{fitted.fire_tensor}
\title{Fitted Values for FIRE Models}
\usage{
\method{fitted}{fire_matrix}(object, ...)

\method{fitted}{fire_tensor}(object, ...)
}
\arguments{
\item{object}{A model object of class \code{fire_matrix} or \code{fire_tensor}.}

\item{...}{Not used.}
}
\value{
A list of class \code{fire_fitted} containing:
\itemize{
  \item \code{yhat}: Fitted values
  \item \code{rmse}: Training RMSE
  \item \code{residuals}: Model residuals
  \item \code{intercept}: Intercept value
  \item \code{model}: Reference to the original model object
}
}
\description{
Computes fitted values for FIRE models with either matrix or tensor input data.
}
\examples{
# For matrix input
data(Manure)
mod <- fire(X = Manure$absorp[1:5,], Y = Manure$y$DM[1:5],
            dat_T = list(1:700), control = list(stop.eps = 2, maxiter = 4))
fitted(mod)

# For tensor input
data(Housing)
dat_T <- list(T1 = 1:4, T2 = 1:9)
mod <- fire(X = Housing$X[1:5,,], Y = Housing$y[1:5,2],
            kernels = list(kronecker_delta, kronecker_delta),
            kernels_params = list(NA, NA),
            dat_T = dat_T, control = list(stop.eps = 2, maxiter = 4))
fitted(mod)
}
\seealso{
\code{\link{fire}}
}
