% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_plot.R
\name{plot.fire_fitted}
\alias{plot.fire_fitted}
\title{Plot for Fitted FIRE Models}
\usage{
\method{plot}{fire_fitted}(x, interval = FALSE, var = NULL, ...)
}
\arguments{
\item{x}{A \code{fire_fitted} object.}

\item{interval}{Logical whether to plot a fitted curve with credible band.}

\item{var}{The predictor to visualize when \code{interval = TRUE}.}

\item{...}{Not used}
}
\value{
Returns a single \code{ggplot} object.
}
\description{
Generates residuals vs fitted values plot for training data or plots a 1D slice of the
fitted function with a credible band for a selected predictor, holding all
other predictors at their column means.
}
\examples{
data(Manure)
mod <- fire(X = Manure$absorp[1:5,], Y = Manure$y$DM[1:5],
            dat_T = list(1:700), control = list(stop.eps = 2, maxiter = 4))
fit <- fitted(mod)
plot(fit)
}
\seealso{
\code{\link{fire}}, \code{\link{fitted.fire}}
}
