% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_predict.R
\name{predict.fire}
\alias{predict.fire}
\alias{predict.fire_matrix}
\alias{predict.fire_tensor}
\title{Predict Method for FIRE Models}
\usage{
\method{predict}{fire_matrix}(object, newdata, Ynew = NULL, interval = FALSE, level = 0.05, ...)

\method{predict}{fire_tensor}(object, newdata, Ynew = NULL, interval = FALSE, level = 0.05, ...)
}
\arguments{
\item{object}{A model object of class \code{fire_matrix} or \code{fire_tensor}.}

\item{newdata}{New data for prediction.}

\item{Ynew}{Optional true response values for test set evaluation.}

\item{...}{Not used.}
}
\value{
A list of class \code{fire_prediction} containing:
\itemize{
  \item \code{yhat}: Predicted values
  \item \code{y.actual}: True values (if \code{Ynew} provided)
  \item \code{test_metrics}: RMSE and residuals (if \code{Ynew} provided)
  \item \code{model}: Reference to the original model
}
}
\description{
Obtain predictions from FIRE models for either matrix or tensor input data.
}
\examples{
# For matrix input
data(Manure)
mod <- fire(X = Manure$absorp[1:5, ], Y = Manure$y$DM[1:5],
            dat_T = list(1:700), control = list(stop.eps = 2, maxiter = 4))
predict(mod, newdata = Manure$absorp[6:10, ])

# For tensor input
data(Housing)
dat_T <- list(T1 = 1:4, T2 = 1:9)
mod <- fire(X = Housing$X[1:5,,], Y = Housing$y[1:5,2],
            kernels = list(kronecker_delta, kronecker_delta),
            kernels_params = list(NA, NA),
            dat_T = dat_T, control = list(stop.eps = 2, maxiter = 4))
predict(mod, newdata = Housing$X[6:10,,])
}
\seealso{
\code{\link{fire}}
}
