% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_predict.R
\name{predict.fire_matrix}
\alias{predict.fire_matrix}
\alias{predict.fire_tensor}
\alias{print.fire_prediction}
\title{Prediction methods for FIRe models}
\usage{
\method{predict}{fire_matrix}(object, newdata, Ynew = NULL, ...)

\method{predict}{fire_tensor}(object, newdata, Ynew = NULL, ...)

\method{print}{fire_prediction}(x, ...)
}
\arguments{
\item{object}{A \code{fire_matrix} or \code{fire_tensor} object}

\item{newdata}{New data for prediction}

\item{Ynew}{Optional numeric vector of true response values for the \code{newdata}. Must be the same length as the number of rows in \code{newdata}.}

\item{...}{Not used}

\item{x}{A \code{fire_prediction} object to print}
}
\value{
A list of class \code{fire_prediction} containing the predicted values,
and including test RMSE, residuals if \code{Ynew} is provided.

The returned object has a \code{print} method that displays:
\itemize{
  \item Number of predictions
  \item Test RMSE (if available)
  \item First few predicted values
}
}
\description{
Obtain predicted values from \code{fire_matrix} or \code{fire_tensor} object, optionally with test set evaluation.
}
\examples{
\dontrun{
data(Manure)
idx <- 1:200
mod <- fire(X = Manure$absorp[idx,], Y = Manure$y$DM[idx], dat_T = list(1:700))
predict(mod, newdata = Manure$absorp[-idx,], Ynew = Manure$y$DM[-idx])
}

}
\seealso{
\code{\link{fire}}
}
