% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_predict.R
\name{predict}
\alias{predict}
\title{Prediction for FIRE Models}
\usage{
predict(object, ...)
}
\arguments{
\item{object}{A model object of class \code{fire_matrix} or \code{fire_tensor}.}

\item{...}{Not used.}

\item{newdata}{New data for prediction.}

\item{Ynew}{Optional true response values for test set evaluation.}
}
\value{
A list of class `fire_prediction` containing:
\itemize{
  \item `yhat`: Predicted values
  \item `y.actual`: True values (if `Ynew` provided)
  \item `test_metrics`: RMSE and residuals (if `Ynew` provided)
  \item `model`: Reference to the original model
}
}
\description{
Obtain predictions from FIRE models of class \code{fire_matrix} or \code{fire_tensor}.
}
\seealso{
\code{\link{fire}}
}
