% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_print.R
\name{print.fire}
\alias{print.fire}
\alias{print.fire_matrix}
\alias{print.fire_tensor}
\title{Print method for FIRe models}
\usage{
\method{print}{fire_matrix}(x, ...)

\method{print}{fire_tensor}(x, ...)
}
\arguments{
\item{x}{A \code{fire_matrix} or \code{fire_tensor} object}

\item{...}{Not used}
}
\value{
The input object (invisibly) for piping. Prints to console:
\itemize{
\item{Basic model information}
\item{Convergence status}
\item{Dimensions}
\item{Estimated hyperparameters}
\item{Marginal log-likelihood}
}
}
\description{
Compact display of \code{fire_matrix} or \code{fire_tensor} objects.
}
\examples{
\dontrun{
data(Manure)
mod <- fire(X = Manure$absorp, Y = Manure$y$DM, dat_T = list(1:700))
print(mod)
}

}
\seealso{
\code{\link{fire}}
}
