% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_predict.R
\name{print.fire_prediction}
\alias{print.fire_prediction}
\title{Print FIRE Predictions}
\usage{
\method{print}{fire_prediction}(x, ...)
}
\arguments{
\item{x}{A \code{fire_prediction} object.}

\item{...}{Not used.}
}
\value{
Invisibly returns the input object.
}
\description{
Summarizes prediction results for \code{fire_prediction} objects.
}
\seealso{
\code{\link{predict.fire}}
}
